requirements.txt
python-telegram-bot==21.2
requests==2.32.3
aiohttp==3.10.10
curl_cffi==0.7.4